Move swagger to separate package to avoid bloating downstream consumers
